Stays Watcher

+++ packages/watcher/migrations/20220418161047-create-block-number.js
@@ -9,7 +9,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       block_number: {
-        type: Sequelize.BIGINT
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
